feat: turn server discovery helpers into scripts
